Update App.js to functional component
