+++ src/components/App.js
@@ -1,33 +1,30 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 
-class App extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = { posts: [] };
+const App = (props) => {
+     const [data, setData] = useState({ posts: [] });
 
-    fetch("http://localhost:4000/posts")
-      .then((response) => response.json())
-      .then((posts) => this.setState({ posts: posts }));
-  }
+     useEffect(() => {
+         fetch("http://localhost:4000/posts")
+           .then((response) => response.json())
+           .then((posts) => setData({ posts: posts }));
+     }, [setData]);
 
-  render() {
-    return (
-      <div>
-        <h1>Hello World</h1>
-        <ul>
-          {this.state.posts.map((post) => {
-            return (
-              <li>
-                <h2>ID: {post.id}</h2>
-                <p>Username: {post.username}</p>
-                <p>Password: {post.password}</p>
-              </li>
-            );
-          })}
-        </ul>
-      </div>
-    );
-  }
-}
+     return (
+       <div>
+         <h1>Hello World</h1>
+         <ul>
+           {data.posts.map((post) => {
+             return (
+               <li>
+                 <h2>ID: {post.id}</h2>
+                 <p>Username: {post.username}</p>
+                 <p>Password: {post.password}</p>
+               </li>
+             );
+           })}
+         </ul>
+       </div>
+     );
+ }
 
 export default App;
